requirements.txt
opendartreader
finance-datareader
polars
psycopg2-binary
requests
python-dotenv
